Tidying up
Remove all experimental concurrency features
Split concurrency tests into separate files
Use group.next() in tests
Revert GH actions to main branch versions
