+++ .github/workflows/benchmark.yml
@@ -1,9 +1,9 @@
 name: Benchmark
 
 on:
-#  pull_request:
-#    branches:
-#    - main
+  pull_request:
+    branches:
+    - main
   workflow_dispatch:
 
 jobs:
