+++ Sources/SotoCore/AWSClient+async.swift
@@ -14,6 +14,7 @@
 
 #if compiler(>=5.5) && $AsyncAwait
 
+import _Concurrency
 import Dispatch
 import Foundation
 import Logging
