+++ Sources/SotoCore/EventLoopFuture+async.swift
@@ -13,6 +13,8 @@
 //===----------------------------------------------------------------------===//
 
 #if compiler(>=5.5) && $AsyncAwait
+
+import _Concurrency
 import NIO
 
 extension EventLoopFuture {
