+++ Tests/SotoCoreTests/AWSServiceTests.swift
@@ -78,48 +78,4 @@ class AWSServiceTests: XCTestCase {
         XCTAssertEqual(service.region, service2.region)
         XCTAssertEqual(service.endpoint, service2.endpoint)
     }
-
-    #if compiler(>=5.5) && $AsyncAwait
-
-    @available(macOS 9999, iOS 9999, watchOS 9999, tvOS 9999, *)
-    func testSignURL() throws {
-        XCTRunAsyncAndBlock {
-            let client = createAWSClient(credentialProvider: .static(accessKeyId: "foo", secretAccessKey: "bar"))
-            defer { XCTAssertNoThrow(try client.syncShutdown()) }
-            let serviceConfig = createServiceConfig()
-            let service = TestService(client: client, config: serviceConfig)
-            let url = URL(string: "https://test.amazonaws.com?test2=true&space=sp%20ace&percent=addi+tion")!
-            let signedURL = try await service.signURL(url: url, httpMethod: .GET, expires: .minutes(15))
-            // remove signed query params
-            let query = try XCTUnwrap(signedURL.query)
-            let queryItems = query
-                .split(separator: "&")
-                .compactMap {
-                    guard !$0.hasPrefix("X-Amz") else { return nil }
-                    return String($0)
-                }
-                .joined(separator: "&")
-            XCTAssertEqual(queryItems, "percent=addi%2Btion&space=sp%20ace&test2=true")
-        }
-    }
-
-    @available(macOS 9999, iOS 9999, watchOS 9999, tvOS 9999, *)
-    func testSignHeaders() throws {
-        XCTRunAsyncAndBlock {
-            let client = createAWSClient(credentialProvider: .static(accessKeyId: "foo", secretAccessKey: "bar"))
-            defer { XCTAssertNoThrow(try client.syncShutdown()) }
-            let serviceConfig = createServiceConfig()
-            let service = TestService(client: client, config: serviceConfig)
-            let url = URL(string: "https://test.amazonaws.com?test2=true&space=sp%20ace&percent=addi+tion")!
-            let headers = try await service.signHeaders(
-                url: url,
-                httpMethod: .GET,
-                headers: ["Content-Type": "application/json"],
-                body: .string("Test payload")
-            )
-            // remove signed query params
-            XCTAssertNotNil(headers["Authorization"].first)
-        }
-    }
-    #endif // compiler(>=5.5) && $AsyncAwait
 }
