+++ Tests/SotoCoreTests/PaginateTests.swift
@@ -334,116 +334,3 @@ class PaginateTests: XCTestCase {
         XCTAssertNoThrow(try future.wait())
     }
 }
-
-#if compiler(>=5.5) && $AsyncAwait
-
-@available(macOS 9999, iOS 9999, watchOS 9999, tvOS 9999, *)
-extension PaginateTests {
-    func counter(_ input: CounterInput, logger: Logger, on eventLoop: EventLoop?) async throws -> CounterOutput {
-        return try await self.counter(input, logger: logger, on: eventLoop).get()
-    }
-
-    func asyncCounterPaginator(_ input: CounterInput) -> AWSClient.PaginatorSequence<CounterInput, CounterOutput> {
-        return .init(
-            input: input,
-            command: self.counter,
-            inputKey: \CounterInput.inputToken,
-            outputKey: \CounterOutput.outputToken,
-            logger: TestEnvironment.logger
-        )
-    }
-
-    func testAsyncIntegerTokenPaginate() throws {
-        XCTRunAsyncAndBlock {
-            // paginate input
-            let input = CounterInput(inputToken: nil, pageSize: 4)
-            async let asyncFinalArray: [Int] = self.asyncCounterPaginator(input).reduce([]) { return $0 + $1.array }
-
-            let arraySize = 23
-            // aws server process
-            XCTAssertNoThrow(try self.awsServer.process { (input: CounterInput) throws -> AWSTestServer.Result<CounterOutput> in
-                // send part of array of numbers based on input startIndex and pageSize
-                let startIndex = input.inputToken ?? 0
-                let endIndex = min(startIndex + input.pageSize, arraySize)
-                var array: [Int] = []
-                for i in startIndex..<endIndex {
-                    array.append(i)
-                }
-                let continueProcessing = (endIndex != arraySize)
-                let output = CounterOutput(array: array, outputToken: endIndex != arraySize ? endIndex : nil)
-                return .result(output, continueProcessing: continueProcessing)
-            })
-
-            let finalArray = try await asyncFinalArray
-            // verify contents of array
-            XCTAssertEqual(finalArray.count, arraySize)
-            for i in 0..<finalArray.count {
-                XCTAssertEqual(finalArray[i], i)
-            }
-        }
-    }
-
-    func testAsyncStringTokenReducePaginate() throws {
-        XCTRunAsyncAndBlock {
-            // paginate input
-            let input = StringListInput(inputToken: nil, pageSize: 5)
-            let paginator = self.asyncStringListPaginator(input)
-            async let asyncResult = paginator.reduce([]) { $0 + $1.array }
-
-            // aws server process
-            XCTAssertNoThrow(try self.awsServer.process(self.stringListServerProcess))
-
-            // wait for response
-            let finalArray = try await asyncResult
-            // verify contents of array
-            XCTAssertEqual(finalArray.count, self.stringList.count)
-            for i in 0..<finalArray.count {
-                XCTAssertEqual(finalArray[i], self.stringList[i])
-            }
-        }
-    }
-
-    func testAsyncPaginateError() throws {
-        XCTRunAsyncAndBlock {
-            // paginate input
-            let input = StringListInput(inputToken: nil, pageSize: 5)
-            let paginator = self.asyncStringListPaginator(input)
-            async let asyncResult = paginator.reduce([]) { $0 + $1.array }
-
-            // aws server process
-            XCTAssertNoThrow(try self.awsServer.process { (_: StringListInput) -> AWSTestServer.Result<StringListOutput> in
-                return .error(.badRequest)
-            })
-
-            do {
-                _ = try await asyncResult
-            } catch {
-                XCTAssertEqual((error as? AWSResponseError)?.errorCode, "BadRequest")
-            }
-        }
-    }
-
-    func stringList(_ input: StringListInput, logger: Logger, on eventLoop: EventLoop? = nil) async throws -> StringListOutput {
-        return try await self.client.execute(
-            operation: "TestOperation",
-            path: "/",
-            httpMethod: .POST,
-            serviceConfig: self.config,
-            input: input,
-            logger: logger,
-            on: eventLoop
-        )
-    }
-
-    func asyncStringListPaginator(_ input: StringListInput) -> AWSClient.PaginatorSequence<StringListInput, StringListOutput> {
-        .init(
-            input: input,
-            command: self.stringList,
-            inputKey: \StringListInput.inputToken,
-            outputKey: \StringListOutput.outputToken,
-            logger: TestEnvironment.logger
-        )
-    }
-}
-
-#endif // compiler(>=5.5) && $AsyncAwait
